incorporated debug mode and shared headers throughout, ready for a test

+++ release/detect_itest_exes.bash
@@ -1,5 +1,10 @@
 #!/bin/bash
 #The purpose of this script is to delete integration tests from release packages.  It deletes tests that rely on unreleased code.  DO NOT RUN THIS SCRIPT EVER (except if you are preparing a release).  Run it from inside tests/integration.
+#author: Steven Lewis, smlewi@gmail.com
+#intended to be run only on Contador, but can be safely edited for whatever other machine is used to create the weekly Rosetta release.
+
+source ./release_common_functions.bash
+
 
 for directory in `ls tests`
 do
@@ -17,7 +22,12 @@ do
 	#if the executable is not found in the released build system, delete the test (as it won't run anyway)
 	echo $executable "not found, REMOVE" $directory
 	git rm -r tests/$directory #--dry-run
-	rm -r ref/$directory
+	if [ "$debug" = false ];
+	    then
+	    rm -r ref/$directory
+	    else
+	    echo "DEBUG MODE ACTIVATED: skipping filesystem ref deletion of autoremoved integration tests"
+	fi
     fi
 done
